requirements.txt
fastapi
uvicorn
pydantic
python-jose
passlib[bcrypt]
langchain
langchain-community
tiktoken
pypdf
sentence-transformers
faiss-cpu
python-dotenv
sqlalchemy
psycopg2-binary 